+++ src/main/java/rip/shukaaa/ui/logic/menu/items/file/ExitMenuItem.java
@@ -1,6 +1,6 @@
 package rip.shukaaa.ui.logic.menu.items.file;
 
-import rip.shukaaa.Main;
+import rip.shukaaa.image.ImageManager;
 import rip.shukaaa.ui.logic.menu.items.MenuItem;
 
 import javax.swing.*;
@@ -10,7 +10,7 @@ public class ExitMenuItem extends MenuItem {
 		protected JMenuItem createItem() {
 				JMenuItem exit = new JMenuItem("Exit");
 				exit.addActionListener(e -> {
-						Main.deleteTempImages();
+						ImageManager.deleteTempImages();
 						System.exit(0);
 				});
 				return exit;
