+++ src/main/java/rip/shukaaa/ui/logic/menu/items/image/ResizeMenuItem.java
@@ -1,6 +1,6 @@
 package rip.shukaaa.ui.logic.menu.items.image;
 
-import rip.shukaaa.Main;
+import rip.shukaaa.image.ImageManager;
 import rip.shukaaa.image.ShukaaaImage;
 import rip.shukaaa.stores.DataStore;
 import rip.shukaaa.ui.logic.menu.items.MenuItem;
@@ -24,7 +24,7 @@ public class ResizeMenuItem extends MenuItem {
 						resizeDialog.add(apply);
 						apply.addActionListener(e1 -> {
 								ShukaaaImage img = DataStore.getImg().resizedImage((int) widthSpinner.getValue(), (int) heightSpinner.getValue());
-								Main.updateImage(img);
+								ImageManager.updateImage(img);
 								resizeDialog.dispose();
 						});
 						resizeDialog.pack();
